Redefined the pipeline object.
To avoid introduction of the 'current' or 'latest' pipeline
versions and implicit handling of the versions, the pipeline
name becomes the unique identifier of the pipeline, with the
uri and version being compulsory attributes on the database
level, but not on the model level.

Client requests can omit the uri and version when claiming
and updating tasks, but are needed when the pipeline record
is created.