Update Task model validation to prevent status being NULL
Change the eq method of Task so that it doesn't raise ValidationError
when the "other" object is a dict which isn't a valid task. It should
simply return that they are not equal.

Simplified the eq method a bit by removing redundant "else"s.

Some expected test results needed changing to accommodate the new Task
status.