Remove default value for status from Task
Avoid setting a default value for status to ensure that it must be set
explicitly. The create_task method in the database layer now has
exclusive control over whether a default (of PENDING) is set, which it
does only when a new task record is created, not when an exisiting
task record is updated.